cmake-mode.el: Drop support for Emacs < 24
The `cmake--parent-mode` alias exists only for compatibility and breaks
Emacs packages and scripts that make use of `prog-mode-hook` because
Emacs does not examine whether the parent mode is an alias or not.
Remove the alias and require Emacs 24 or higher.

Emacs 24 was released in 2012 and this is already 2019, thus the loss to
drop support for versions < 24 is negligible.